First Commit
This is the first commit of the Otis-Frontend project. Basic application functionality was established for now and setup for future additions to the code. Redux store, base components, root reducer, router, are ready for future code implementation.

Basic testing was also added for existing components and devServer.

Hot realoding for development was added and a seperate Dev/Production Webpack config is ready for future use.